+++ NEWS.md
@@ -4,6 +4,7 @@ CoordinateCleaner 2.0-7 (2019-01-22)
 ### MINOR IMPROVEMENTS
   * added citation
   * reduced testing time on CRAN
+  * improved documentation of the cc_outl function
 
 CoordinateCleaner 2.0-6 (2019-01-16)
 =========================